requirements.txt
sanic==23.12.1
jinja2==3.1.3
sqlalchemy==1.4.51
aiosqlite==0.20.0
gitpython==3.1.41 